# Conflicts:
#	apps/ahrefs/src/config/ahrefs.config.joi.ts
#	apps/ahrefs/src/config/ahrefs.config.ts
#	apps/aws/src/modules/rds/aws.rds.service.ts
#	apps/aws/src/modules/rds/config/aws.rds.config.ts
#	apps/aws/src/modules/rds/index.ts
#	apps/aws/src/modules/s3/aws.s3.service.ts
#	apps/aws/src/modules/secrets/aws.secrets.service.ts
#	apps/aws/src/modules/ses/aws.ses.service.ts
#	apps/aws/src/modules/sns/aws.sns.service.ts
#	apps/everflow/src/modules/affiliate/offers/everflow.affiliate.offers.service.ts
#	apps/everflow/src/modules/affiliate/reporting/clicks/everflow.affiliate.reporting.clicks.service.ts
#	apps/everflow/src/modules/affiliate/reporting/conversions/everflow.affiliate.reporting.conversions.service.ts
#	apps/openai/src/module/openai.service.ts
#	apps/pexels/src/module/pexels.service.ts
#	apps/scrapingbee/src/sandbox/main.ts
#	apps/scrapingbee/src/sandbox/sandbox.controller.ts
#	apps/scrapingbee/src/scrapingbee/scrapingbee.scrape.service.ts
#	apps/wordpress/package.json
#	apps/wordpress/src/modules/media/wordpress.media.service.test.spec.ts
#	apps/wordpress/src/modules/posts/wordpress.posts.service.ts
#	apps/wordpress/src/sandbox/main.ts
#	apps/xero-cc/src/module/auth/auth.service.ts
#	apps/xero-cc/src/module/customer/contact.entity.ts
#	apps/xero-cc/src/module/customer/contact.service.ts
#	apps/xero-cc/src/module/invoice/invoice.entity.ts
#	apps/xero-cc/src/module/invoice/invoice.service.ts
#	apps/xero-cc/src/module/xero.webhook.controller.ts
#	backend/core/src/helpers/crudController.ts
#	backend/core/src/modules/auth/auth.module.ts
#	backend/core/src/test/helpers.ts
#	backend/core/src/test/scaffold.ts
#	pnpm-lock.yaml

+++ apps/wordpress/src/modules/media/wordpress.media.service.ts
@@ -32,6 +32,7 @@ export class WordpressMediaService {
 			this.logger.debug(`[${domain}] Skipping as in test mode`)
 			return <any>mock
 		}
+		if (!options.config) throw new Error('Missing config')
 
 		try {
 			const url = new URL(this.urlBase)
@@ -45,8 +46,10 @@ export class WordpressMediaService {
 			})
 
 			return await this.update({ mediaId: media.id, data: options.data, config: options.config })
-		} catch (e) {
+		} catch (err) {
+			const e = err as Error
 			this.logger.error(`[${domain}] Error creating media: ${e.message}`, e)
+			throw e
 		}
 	}
 
@@ -60,13 +63,16 @@ export class WordpressMediaService {
 			this.logger.debug(`[${domain}] Skipping as in test mode`)
 			return [<any>mock]
 		}
+		if (!options?.config) throw new Error('Missing config')
 
 		try {
 			const url = new URL(this.urlBase)
 			url.search = new URLSearchParams(<any>options.arguments).toString()
 			return await this.api.get(domain, url.toString())
-		} catch (e) {
+		} catch (err) {
+			const e = err as Error
 			this.logger.error(`[${domain}] Error finding all media: ${e.message}`, e)
+			throw e
 		}
 	}
 
@@ -81,13 +87,16 @@ export class WordpressMediaService {
 			this.logger.debug(`[${domain}] Skipping as in test mode`)
 			return <any>mock
 		}
+		if (!options.config) throw new Error('Missing config')
 
 		try {
 			const url = new URL(`${this.urlBase}/${options.mediaId}`)
 			url.search = new URLSearchParams(<any>options.arguments).toString()
 			return await this.api.get(domain, url.toString(), getWordpressAxiosConfig(options.config))
-		} catch (e) {
+		} catch (err) {
+			const e = err as Error
 			this.logger.error(`[${domain}] Error finding one media: ${e.message}`, e)
+			throw e
 		}
 	}
 
@@ -102,12 +111,15 @@ export class WordpressMediaService {
 			this.logger.debug(`[${domain}] Skipping as in test mode`)
 			return <any>mock
 		}
+		if (!options.config) throw new Error('Missing config')
 
 		try {
 			const url = new URL(`${this.urlBase}/${options.mediaId}`)
 			return await this.api.post(domain, url.toString(), options.data, getWordpressAxiosConfig(options.config))
-		} catch (e) {
+		} catch (err) {
+			const e = err as Error
 			this.logger.error(`[${domain}] Error updating media: ${e.message}`, e)
+			throw e
 		}
 	}
 
@@ -118,12 +130,15 @@ export class WordpressMediaService {
 			this.logger.debug(`[${domain}] Skipping as in test mode`)
 			return
 		}
+		if (!options.config) throw new Error('Missing config')
 
 		try {
 			const url = new URL(`${this.urlBase}/${options.mediaId}`)
 			await this.api.delete(domain, url.toString(), getWordpressAxiosConfig(options.config))
-		} catch (e) {
+		} catch (err) {
+			const e = err as Error
 			this.logger.error(`[${domain}] Error deleting media: ${e.message}`, e)
+			throw e
 		}
 	}
 }
